refactor(wallet): shrink to network-only shim; split BIP-380 checksum helper
