+++ main/core/descriptor_checksum.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <stdbool.h>
+
+bool wallet_get_descriptor_string(char **output);
+bool wallet_get_descriptor_checksum(char **output);
