+++ main/core/test/stubs/esp_err.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <stdint.h>
-typedef int32_t esp_err_t;
+typedef int esp_err_t;
 #define ESP_OK 0
