+++ main/pages/home/public_key.c
@@ -47,11 +47,11 @@ void public_key_page_create(lv_obj_t *parent, void (*return_cb)(void)) {
 
   return_callback = return_cb;
 
-  const char *derivation_path = wallet_get_derivation();
+  uint32_t coin = (wallet_get_network() == WALLET_NETWORK_MAINNET) ? 0 : 1;
+  char derivation_path[48];
+  snprintf(derivation_path, sizeof(derivation_path), "m/84'/%u'/0'", coin);
   char derivation_compact[32];
-  wallet_format_derivation_compact(
-      derivation_compact, sizeof(derivation_compact), wallet_get_policy(),
-      wallet_get_network(), wallet_get_account());
+  snprintf(derivation_compact, sizeof(derivation_compact), "84h/%uh/0h", coin);
 
   public_key_screen = lv_obj_create(parent);
   lv_obj_set_size(public_key_screen, LV_PCT(100), LV_PCT(100));
